+++ front/src/app/questions/question-list/question-list.component.ts
@@ -2,6 +2,8 @@ import {Component, Input, OnInit} from '@angular/core';
 import {Quiz} from '../../../models/quiz.model';
 import {QuizService} from '../../../services/quiz.service';
 import {Question} from '../../../models/question.model';
+import {PopUpVerifComponent} from '../../patients/patient-profile/patient-infos/PopupVerif/pop-up.component';
+import {MatDialog} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-question-list',
@@ -13,13 +15,28 @@ export class QuestionListComponent implements OnInit {
   @Input()
   quiz: Quiz;
 
-  constructor(private quizService: QuizService) { }
+  private result;
+
+  constructor(public dialog: MatDialog, private quizService: QuizService) { }
 
   ngOnInit() {
   }
 
   deleteQuestion(question: Question) {
-    this.quizService.deleteAnswers(this.quiz, question);
-    this.quizService.deleteQuestion(this.quiz, question);
+    if (this.result) {
+      this.quizService.deleteAnswers(this.quiz, question);
+      this.quizService.deleteQuestion(this.quiz, question);
+    }
+  }
+
+  openPop(question: Question): void {
+    const dialogRef = this.dialog.open(PopUpVerifComponent, {
+      width: '250px',
+      data: { name: 'la question'}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      this.result = result;
+      this.deleteQuestion(question);
+    });
   }
 }
